project/requirements.txt
pandas>=1.3.0
tqdm>=4.60.0
Pillow>=8.0.0
scikit-learn>=0.24.0
PyYAML>=5.4.0